+++ src/main/java/com/github/siroshun09/sirolibrary/SiroLibraryBungee.java
@@ -17,6 +17,7 @@ public class SiroLibraryBungee extends Plugin {
      */
     public SiroLibraryBungee() {
         instance = this;
+        SiroLibraryLogger.setLogger(getLogger());
     }
 
     /**