+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # 変更履歴
 
+## 1.4.7
+
+- BukkitUtil - 権限系メソッド
+- FileLogger - 定数名を大文字化
+- FileLogger - addDate(String, boolean) の追加 (改行の有無)
+- FileLogger - addDate() での日付フォーマット変更
+- WriteTask - Files.isWritable で書き込みチェック
+- SiroLibraryLogger - 稼働場所に関わらず、ロガーを取得する。
+- Guava 削除 (使用しないため)
+
 ## 1.4.6
 
 - NumberChecker - Int または Long であるか判定するメソッド